Moved source files to subfolder and updated imports.
